requirements.txt
markdownify
smolagents
requests
duckduckgo_search
pandas
smolagents[gradio]
python-dotenv
simple-salesforce

# huggingface_hub